+++ src/finance/programgarden_finance/__init__.py
@@ -87,6 +87,7 @@ from .ls.korea_stock.program import t1632
 from .ls.korea_stock.program import t1633
 from .ls.korea_stock.program import t1636
 from .ls.korea_stock.program import t1637
+from .ls.korea_stock.program import t1640
 from .ls.korea_stock.sector import t1511
 from .ls.korea_stock.sector import t1516
 from .ls.korea_stock.sector import t1531
@@ -193,6 +194,7 @@ __all__ = [
     t1633,
     t1636,
     t1637,
+    t1640,
 
     COSAQ00102,
     COSAQ01400,